Migrating project to systemjs-builder
