+++ client/index.html
@@ -6,9 +6,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width; initial-scale=1.0; maximum-scale=1.0; user-scalable=0;" />
     <link rel="stylesheet" href="style.css">
-    <script src="../node_modules/handlebars/dist/handlebars.js"></script>
-    <script src="../node_modules/systemjs/dist/system.js"></script>
-    <script src="system.config.js"></script>
+    <script src="node_modules/handlebars/dist/handlebars.js"></script>
+    <script src="dist/app.js" defer></script>
   </head>
 
   <body>
