+++ package.json
@@ -9,22 +9,21 @@
   "dependencies": {
     "express": "*",
     "body-parser": "*",
-    "mongoose": "*"
+    "mongoose": "*",
+    "systemjs-builder": "^0.5.3"
   },
   "devDependencies": {
-    "babel-core": "^6.7.7",
-    "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.6.0",
-    "babelify": "~5.0.3",
-    "browser-sync": "^2.11.1",
-    "browserify": "^13.0.0",
-    "es6-module-loader": "^0.17.11",
+    "babelify": "^7.3.0",
+    "browserify": "^13.0.1",
     "gulp": "^3.9.1",
     "gulp-autoprefixer": "^3.1.0",
-    "gulp-babel": "^6.1.2",
+    "gulp-babel": "^4.0.0",
+    "gulp-concat": "^2.4.3",
     "gulp-exec": "^2.1.2",
     "gulp-nodemon": "^2.0.6",
     "gulp-sass": "^2.2.0",
+    "gulp-watch": "^4.1.0",
     "handlebars": "^4.0.5",
     "jasmine-core": "^2.4.1",
     "karma": "^0.13.22",
@@ -34,8 +33,7 @@
     "karma-jasmine": "^0.3.8",
     "karma-safari-launcher": "^0.1.1",
     "karma-systemjs": "^0.13.0",
-    "systemjs": "^0.19.26",
-    "webpack": "^1.13.0"
+    "run-sequence": "^1.0.2"
   },
   "author": "",
   "license": "ISC"
